pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "eia-gen"
version = "0.1.0"
description = "Small-scale EIA (tourist farm) DOCX generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "openpyxl>=3.1.5",
  "pydantic>=2.7.4",
  "pydantic-settings>=2.4.0",
  "Pillow>=10.4.0",
  "pyproj>=3.6.1",
  "python-multipart>=0.0.9",
  "PyYAML>=6.0.2",
  "python-docx>=1.1.2",
  "rich>=13.8.1",
  "shapely>=2.0.6",
  "typer>=0.12.5",
  "uvicorn[standard]>=0.30.6",
]

[project.scripts]
eia-gen = "eia_gen.cli:app"

[project.optional-dependencies]
# Optional tools for "sample-style" image processing utilities (not required for core docx generation).
image-tools = [
  "scikit-image>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/eia_gen"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
